Don't stop spinner too early in GitLabClient's selectRepo
We don't want to stop the spinner early, while repos are still loading
in.

I've only updated selectRepo, because that's the only one we're having
issues with. In theory, a similar issue probably exists with selectFile
and selectRef, but the number of files/refs are likely to be so small
that it won't be a problem in practice.